1.1.3 修復只有在第一頁點選上一頁才跳出提示 修復評分出現浮點數時呼叫API格式錯誤 修復顯示是只顯示小數前一位

+++ app/src/main/java/com/example/idiotchefassistant/recipe/RecipeItem.kt
@@ -8,7 +8,7 @@ data class RecipeItem(
     val rid:Int,
     val title:String,
     val link:String,
-    val score:Int
+    val score:Float
 ): Parcelable
 
 //data class CommitItem(val userImg: Int, val userName:String, val context:String, val review:Int)
